requirements.txt
requests
streamlit
python-dotenv
pandas
plotly
openai
numpy
scikit-learn